binder: Add readuInt64() symbol for LG RIL compatibility
 * LG RIL libs need readuInt64() (I have no idea why, blame LG)
 * Instead of having to use liblgderp.so we can just set
   NEEDS_LGE_RIL_SYMBOLS to incorporate the symbols into the system
 * Original commits:
   - http://review.cyanogenmod.org/#/c/61459/
   - http://review.cyanogenmod.org/#/c/61460/

+++ include/binder/Parcel.h
@@ -155,7 +155,7 @@ public:
     int64_t             readInt64() const;
     status_t            readInt64(int64_t *pArg) const;
 #ifdef NEEDS_LGE_RIL_SYMBOLS
-    status_t            readuInt64(unsigned long long *pArg) const;
+    status_t            readuInt64(uint64_t *pArg) const;
 #endif
     float               readFloat() const;
     status_t            readFloat(float *pArg) const;
